# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Eval API
  description: |
    AI Agent Safety Evaluation Platform API.

    This API allows you to:
    - Register AI agents for evaluation
    - Run safety evaluations against your agents (red_team and trust)
    - Evaluate agentic AI systems with tool access
    - View detailed results, ASR metrics, and recommendations

    ## Authentication

    All endpoints require authentication via:
    - **Bearer Token**: Pass `Authorization: Bearer <api_key>` header

    ## Evaluation Types

    | Type | Purpose | Tests |
    |------|---------|-------|
    | red_team | Security testing | Jailbreaks, injection attacks, data exfiltration |
    | trust | Quality testing | Hallucination, bias, over-refusal |

    ## Evaluation Modes

    | Mode | Duration | Prompts | Use Case |
    |------|----------|---------|----------|
    | quick | ~5 min | ~100 | Smoke test, CI/CD |
    | standard | ~30 min | ~2,000 | Regular testing |
    | deep | ~2+ hrs | ~5,000+ | Comprehensive audit |

    ## Agentic AI Evaluation

    For AI agents with tool access (Salesforce, Jira, Slack, etc.), we provide:
    - **ASR Metrics**: Attack Success Rate (full/partial/blocked)
    - **Multi-turn Attack Detection**: Escalating conversation attacks that build trust before exploitation
    - **Memory/Context Attacks**: Tests for RAG poisoning and context manipulation
    - **Tool-aware Scenarios**: Custom scenarios targeting specific integrations
...

    ## Agentic Attack Categories
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents:
    parameters:
    get:
      tags: [Agents]
      summary: List agents
      description: List all registered agents for the organization
      operationId: listAgents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, deleted]
          description: Filter by agent status
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Agents]
      summary: Register agent
      description: |
        Register a new AI agent for evaluation.

        ## Supported Platforms

        | Protocol | Required Fields | Auth Types |
        |----------|-----------------|------------|
        | `openai` | promptId OR assistantId OR workflowId | bearer (API key) |
        | `azure_openai` | baseUrl, promptId OR assistantId OR workflowId | azure_entra, api_key |
        | `langgraph` | baseUrl, assistantId | bearer |
        | `bedrock` | agentId | aws |
        | `vertex` | projectId, reasoningEngineId | google_oauth |
        | `custom` | endpoint | any |
        | `a2a` | agentCardUrl | any |
        | `mcp` | endpoint, mcpTransport | any |

        ## Platform Protocols

        For agents on major platforms, we use native SDKs for optimal compatibility:
        - **OpenAI**: Supports Responses API (promptId), Assistants API (assistantId, deprecated Aug 2026), and Agent Builder (workflowId)
        - **Azure OpenAI**: Same as OpenAI with Azure Entra ID auth
        - **LangGraph**: Uses threads/runs model via langgraph-sdk
        - **Bedrock**: Uses AWS Bedrock Agents via boto3
        - **Vertex**: Uses Google Vertex AI Agent Engine

        ## Custom Agents

        For custom HTTP APIs, use protocol `custom` and provide:
        - `endpoint`: Full URL to your conversation endpoint (e.g., `https://my-agent.example.com/api/v1/chat`)

        We'll automatically analyze your API to infer the request/response format.

        ## Future Protocols (V2)

        - **A2A**: Google Agent-to-Agent protocol - requires `agentCardUrl`
        - **MCP**: Model Context Protocol - requires `endpoint` and `mcpTransport`
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              openaiResponses:
                summary: OpenAI Prompt (Responses API)
                description: Register an OpenAI Prompt using the Responses API (preferred).
                value:
                  agentId: my-openai-agent
                  name: OpenAI Agent
                  description: Customer support agent built on OpenAI
                  goal: Answer customer questions
                  industry: retail
                  endpointConfig:
                    protocol: openai
                    promptId: pmpt_abc123xyz
                  authConfig:
                    type: bearer
                    credentials: sk-xxx
              openaiAssistants:
                summary: OpenAI Assistant (legacy)
                description: Register an OpenAI Assistant using the Assistants API (deprecated Aug 2026).
                value:
                  agentId: my-openai-agent
                  name: OpenAI Assistant
                  description: Customer support assistant built on OpenAI
                  goal: Answer customer questions
                  industry: retail
                  endpointConfig:
                    protocol: openai
                    assistantId: asst_abc123xyz
                  authConfig:
                    type: bearer
                    credentials: sk-xxx
              openaiAgentBuilder:
                summary: OpenAI Agent Builder (ChatKit)
                description: Register an OpenAI Agent Builder workflow using ChatKit.
                value:
                  agentId: my-workflow-agent
                  name: Multi-Agent Workflow
                  description: Multi-agent workflow built with Agent Builder
                  goal: Handle complex multi-step tasks
                  industry: technology
                  endpointConfig:
                    protocol: openai
                    workflowId: wf_abc123xyz
                  authConfig:
                    type: bearer
                    credentials: sk-xxx
              azureOpenai:
                summary: Azure OpenAI
                description: Register an Azure OpenAI agent with Entra ID auth.
                value:
                  agentId: my-azure-agent
                  name: Azure OpenAI Agent
                  description: Enterprise agent hosted on Azure
                  goal: Enterprise customer support
                  industry: finance
                  endpointConfig:
                    protocol: azure_openai
                    baseUrl: https://my-resource.openai.azure.com
                    promptId: pmpt_abc123xyz
                  authConfig:
                    type: azure_entra
                    credentials: client_id:client_secret
                    azureTenantId: 12345678-1234-1234-1234-123456789012
              langgraph:
                summary: LangGraph Platform agent
                description: Register a LangGraph Platform deployed agent.
                value:
                  agentId: my-langgraph-agent
                  name: LangGraph Agent
                  description: Multi-step reasoning agent
                  goal: Complex task automation
                  industry: technology
                  endpointConfig:
                    protocol: langgraph
                    baseUrl: https://my-deployment.langchain.app
                    assistantId: my-assistant
                  authConfig:
                    type: bearer
                    credentials: lsv2_xxx
              bedrock:
                summary: AWS Bedrock agent
                description: Register an AWS Bedrock agent.
                value:
                  agentId: my-bedrock-agent
                  name: Bedrock Agent
                  description: AWS Bedrock powered assistant
                  goal: Enterprise automation
                  industry: healthcare
                  endpointConfig:
                    protocol: bedrock
                    agentId: ABCD1234XYZ
                    agentAliasId: TSTALIASID
                    region: us-east-1
                  authConfig:
                    type: aws
                    credentials: AKIAIOSFODNN7EXAMPLE:wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY
                    awsRegion: us-east-1
              vertex:
                summary: Google Vertex AI agent
                description: Register a Vertex AI Agent Engine agent.
                value:
                  agentId: my-vertex-agent
                  name: Vertex AI Agent
                  description: Google Cloud powered agent
                  goal: Customer service automation
                  industry: retail
                  endpointConfig:
                    protocol: vertex
                    projectId: my-gcp-project
                    region: us-central1
                    reasoningEngineId: "1234567890"
                  authConfig:
                    type: google_oauth
                    credentials: '{"type":"service_account",...}'
              custom:
                summary: Custom HTTP agent
                description: Register a custom HTTP API agent.
                value:
                  agentId: my-custom-agent
                  name: My Custom Agent
                  description: A custom AI assistant
                  goal: Answer customer questions
                  industry: retail
                  endpointConfig:
                    protocol: custom
                    endpoint: https://my-agent.example.com/api/v1/chat
                    specUrl: https://my-agent.example.com/openapi.yaml
                  authConfig:
                    type: api_key
                    credentials: xxx
                    headerName: X-API-Key
              a2a:
                summary: A2A agent (V2)
                description: Register a Google Agent-to-Agent protocol agent.
                value:
                  agentId: my-a2a-agent
                  name: A2A Agent
                  description: Agent using Google A2A protocol
                  goal: Multi-agent collaboration
                  industry: technology
                  endpointConfig:
                    protocol: a2a
                    agentCardUrl: https://my-agent.example.com/.well-known/agent.json
                  authConfig:
                    type: bearer
                    credentials: xxx
              mcp:
                summary: MCP server (V2)
                description: Register an MCP (Model Context Protocol) server.
                value:
                  agentId: my-mcp-agent
                  name: MCP Agent
                  description: Agent using MCP protocol
                  goal: Tool integration
                  industry: technology
                  endpointConfig:
                    protocol: mcp
                    endpoint: https://my-mcp-server.example.com/mcp
                    mcpTransport: sse
                  authConfig:
                    type: bearer
                    credentials: xxx
              withCapabilities:
                summary: Agent with declared capabilities
                description: Register an agent with tool access for agentic evaluation.
                value:
                  agentId: my-agent
                  name: My Assistant
                  description: A helpful AI assistant with tools
                  goal: Answer customer questions
                  industry: retail
                  endpointConfig:
                    protocol: openai
                    promptId: pmpt_abc123xyz
                  authConfig:
                    type: bearer
                    credentials: sk-xxx
                  sessionStrategy: agent_managed
                  declaredCapabilities:
                    hasMemory: true
                    hasTools: true
                    tools: ["salesforce", "slack", "jira"]
                    sensitiveDataTypes: ["pii", "financial"]
                    hasRag: true
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /agents/{agentId}:
    parameters:
      - $ref: '#/components/parameters/agentId'

    get:
      tags: [Agents]
      summary: Get agent
      description: Get details of a specific agent version by snapshotId.
      operationId: getAgent
      parameters:
        - name: snapshotId
          in: query
          required: true
          schema:
            type: string
          description: Snapshot ID to retrieve a specific agent version
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Agents]
      summary: Update agent
      description: Update agent configuration
      operationId: updateAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Agents]
      summary: Delete agent
      description: Soft delete an agent (can be restored)
      operationId: deleteAgent
      responses:
        '204':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/analyze:
    parameters:
      - $ref: '#/components/parameters/agentId'
      - $ref: '#/components/parameters/snapshotId'
    post:
      tags: [Agents]
      summary: Analyze agent API
      description: |
        Use GPT to probe and analyze the agent's API structure.

        This endpoint:
        1. Parses any provided OpenAPI spec
        2. Probes the endpoint with common request formats
        3. Uses GPT to analyze responses and infer the correct schema

        The inferred schema is automatically saved to the agent and used for
        subsequent evaluations.

        **Note**: This runs automatically when an agent is created with HTTP protocol,
        but can be triggered manually to re-analyze after configuration changes.
      operationId: analyzeAgentAPI
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeAPIResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # Scenarios
  # ============================================================================
  /agents/{agentId}/scenarios/generate:
    parameters:
      - $ref: '#/components/parameters/agentId'
    post:
      tags: [Scenarios]
      summary: Generate scenarios
      description: |
        Generate custom attack scenarios tailored to your agent's capabilities.

        ## Eval Types

        - **red_team**: Security testing - prompts designed to trick the agent into unsafe behavior
        - **trust**: Quality testing - prompts to check for hallucination, bias, over-refusal

        ## Focus Areas

        **Red Team - Agentic-Specific (Highest Priority)**:
        These focus areas target vulnerabilities specific to AI agents with tool execution:
        - `memory_poisoning`: Poison agent memory/RAG for later exploitation (multi-turn attacks)
        - `tool_chaining`: Exploit tool combinations for unauthorized actions
        - `goal_hijacking`: Redirect agent objectives via embedded instructions
        - `trust_exploitation`: Build trust over multiple turns then exploit

        **Red Team - Traditional**:
        - `data_exfiltration`: Extract sensitive data through tools
        - `unauthorized_actions`: Perform actions beyond authorization
        - `privilege_escalation`: Gain elevated permissions
        - `credential_theft`: Extract API keys, passwords, tokens
        - `social_engineering`: Psychological manipulation tactics
        - `prompt_injection`: Inject malicious instructions via data
        - `jailbreak`: Bypass safety guidelines
        - `harmful_content`: Generate harmful content

        **Trust** (agent should HELP appropriately):
        - `hallucination`: Invents facts or citations
        - `over_refusal`: Refuses legitimate requests
        - `bias`: Shows demographic bias
        - `privacy_handling`: PII handling
        - `uncertainty`: False confidence
        - `helpfulness`: Actually answers questions
        - `professional_redirect`: Knows when to recommend professionals

        ## Defaults

        **Red Team defaults** (agentic-prioritized): `memory_poisoning`, `tool_chaining`, `data_exfiltration`, `privilege_escalation`, `trust_exploitation`

        **Trust defaults**: `hallucination`, `over_refusal`, `professional_redirect`, `bias`

        ## Multi-Turn Emphasis

        For agentic evaluation, `includeMultiTurn` defaults to **true**. Multi-turn scenarios are critical for testing:
        - Memory/context poisoning attacks
        - Trust building followed by exploitation
        - Tool chaining across conversation turns
        - Gradual privilege escalation

        Generation is asynchronous. Use the streaming endpoint for live progress.
      operationId: generateScenarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateScenariosRequest'
            examples:
              redTeamAgentic:
                summary: Red team with agentic focus areas (recommended)
                value:
                  evalType: red_team
                  tools: ["salesforce", "jira", "slack", "email"]
                  focusAreas: ["indirect_injection", "rag_poisoning", "cascading_failure", "tool_chaining"]
                  count: 25
                  includeMultiTurn: true
              redTeamTraditional:
                summary: Red team with traditional focus areas
                value:
                  evalType: red_team
                  tools: ["salesforce", "jira", "slack"]
                  focusAreas: ["data_exfiltration", "social_engineering", "privilege_escalation"]
                  count: 25
                  includeMultiTurn: true
              redTeamCodeExecution:
                summary: Red team testing code execution risks
                value:
                  evalType: red_team
                  tools: ["code_interpreter", "shell", "file_system"]
                  focusAreas: ["code_execution", "supply_chain", "indirect_injection"]
                  count: 20
                  includeMultiTurn: true
              redTeamMultiAgent:
                summary: Red team for multi-agent systems
                value:
                  evalType: red_team
                  tools: ["agent_orchestrator", "sub_agents", "shared_memory"]
                  focusAreas: ["inter_agent", "cascading_failure", "rag_poisoning"]
                  count: 20
                  includeMultiTurn: true
              redTeamWithContext:
                summary: Red team with uploaded documents
                value:
                  evalType: red_team
                  tools: ["salesforce", "jira", "slack"]
                  focusAreas: ["indirect_injection", "rag_poisoning", "data_exfiltration"]
                  contextIds: ["ctx_20241208_abc123"]
                  customInstructions: "Focus on attacks that leverage the uploaded business context"
                  count: 25
                  includeMultiTurn: true
              trust:
                summary: Trust evaluation scenarios
                value:
                  evalType: trust
                  focusAreas: ["hallucination", "over_refusal", "bias"]
                  customInstructions: "Focus on healthcare compliance scenarios. Test if agent appropriately redirects medical questions to professionals."
                  count: 20
              trustWithContext:
                summary: Trust evaluation with uploaded documents
                value:
                  evalType: trust
                  focusAreas: ["hallucination", "over_refusal", "bias"]
                  contextIds: ["ctx_20241208_abc123", "ctx_20241208_def456"]
                  count: 20
      responses:
        '202':
          description: Generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateScenariosResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/scenarios:
    parameters:
      - $ref: '#/components/parameters/agentId'
    get:
      tags: [Scenarios]
      summary: List scenario sets
      description: List all generated scenario sets for an agent
      operationId: listScenarios
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of scenario sets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioSetListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/scenarios/{scenarioSetId}:
    parameters:
      - $ref: '#/components/parameters/agentId'
      - $ref: '#/components/parameters/scenarioSetId'
    get:
      tags: [Scenarios]
      summary: Get scenario set
      description: Get details of a specific scenario set including all scenarios
      operationId: getScenarioSet
      responses:
        '200':
          description: Scenario set details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioSet'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Scenarios]
      summary: Update scenario set
      description: Enable/disable scenarios or add/remove scenarios from a set
      operationId: updateScenarioSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScenariosRequest'
      responses:
        '200':
          description: Scenario set updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioSet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Scenarios]
      summary: Delete scenario set
      description: Delete a scenario set
      operationId: deleteScenarioSet
      responses:
        '204':
          description: Scenario set deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /contexts:
    parameters:
    post:
      tags: [Contexts]
      summary: Upload context documents
      description: |
        Upload documents (PDF, Markdown, CSV, Excel) to provide additional context
        for scenario generation.

        ## Supported File Types

        - **PDF** (.pdf) - Text extracted from all pages
        - **Markdown** (.md) - Direct text content
        - **CSV** (.csv) - Parsed as tabular data
        - **Excel** (.xlsx, .xls) - Parsed as tabular data
        - **Text** (.txt) - Direct text content
        - **JSON** (.json) - Parsed JSON data

        ## Processing

        Files are uploaded immediately and processed asynchronously:
        1. Upload: Files stored in cloud storage
        2. Processing: Text extraction and parsing
        3. Ready: Context available for scenario generation

        Poll the GET endpoint to check processing status.

        ## Usage in Scenario Generation

        Once a context is ready, reference it by `contextId` when generating scenarios:
        ```json
        {
          "contextIds": ["ctx_20241208_abc123"],
          "evalType": "red_team",
          ...
        }
        ```
      operationId: createContext
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, files]
              properties:
                name:
                  type: string
                  description: Human-readable name for the context
                  example: Healthcare Compliance Docs
                description:
                  type: string
                  description: Optional description
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more files to upload (max 50MB each)
      responses:
        '201':
          description: Context created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: Context service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [Contexts]
      summary: List contexts
      description: List all uploaded contexts for the organization
      operationId: listContexts
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of contexts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextListResponse'
  /contexts/{contextId}:
    parameters:
      - $ref: '#/components/parameters/contextId'
    get:
      tags: [Contexts]
      summary: Get context
      description: Get details of a specific context including file status
      operationId: getContext
      responses:
        '200':
          description: Context details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Contexts]
      summary: Delete context
      description: Delete a context and all its files
      operationId: deleteContext
      responses:
        '204':
          description: Context deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /evals/{runId}:
    parameters:
      - $ref: '#/components/parameters/runId'

    get:
      tags: [Evaluations]
      summary: Get evaluation status
      description: Get the current status and progress of an evaluation run
      operationId: getEval
      responses:
        '200':
          description: Evaluation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRun'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Bearer token authentication (Authorization: Bearer <api_key>)"
  parameters:
...
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Offset for pagination
...